requirements.txt
asn1crypto==0.24.0
attrdict==2.0.0
bytestringSplitter==1.0.0a1
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
constant-sorrow==0.1.0a3
cryptography==2.3.1
cytoolz==0.9.0.1
Django==2.1.2
djangorestframework==3.9.0
eth-abi==1.2.2
eth-account==0.3.0
eth-bloom==1.0.1
eth-hash==0.2.0
eth-keyfile==0.5.1
eth-keys==0.2.0b3
eth-rlp==0.1.2
eth-tester==0.1.0b32
eth-typing==2.0.0
eth-utils==1.3.0b0
hexbytes==0.1.0
idna==2.7
lru-dict==1.1.6
msgpack-python==0.5.6
parsimonious==0.8.1
py-ecc==1.4.3
py-evm==0.2.0a32
py-geth==2.0.1
py-solc==3.2.0
pycparser==2.19
pycryptodome==3.6.6
pyethash==0.1.27
PyNaCl==1.3.0
pytz==2018.5
requests==2.20.0
rlp==1.0.3
semantic-version==2.6.0
six==1.11.0
toolz==0.9.0
trie==1.3.8
umbral==0.1.1a3
urllib3==1.24
web3==4.7.2
websockets==6.0
